MAINT: stats: move entropy and differential_entropy functions to more appropriate file
