+++ scipy/stats/distributions.py
@@ -5,8 +5,7 @@
 # NOTE: To look at history using `git blame`, use `git blame -M -C -C`
 #       instead of `git blame -Lxxx,+x`.
 #
-from ._distn_infrastructure import (entropy, differential_entropy,
-                                    rv_discrete, rv_continuous, rv_frozen)
+from ._distn_infrastructure import (rv_discrete, rv_continuous, rv_frozen)
 
 from . import _continuous_distns
 from . import _discrete_distns
@@ -15,8 +14,7 @@ from ._continuous_distns import *
 from ._discrete_distns import *
 
 # For backwards compatibility e.g. pymc expects distributions.__all__.
-__all__ = ['entropy', 'differential_entropy',
-           'rv_discrete', 'rv_continuous', 'rv_histogram']
+__all__ = ['rv_discrete', 'rv_continuous', 'rv_histogram']
 
 # Add only the distribution names, not the *_gen names.
 __all__ += _continuous_distns._distn_names
